[REFACTOR] : LoginViewModel
